refactor: break out common code from txn test
- common tooling:
  - getting public key address
  - getting node's txn params
  - signing tagged txn
